Major rearrangement of functions to modules
