--- nanometa_live/gui_scripts/__init__.py
@@ -1,1 +1,0 @@
-name = "gui_scripts"
